nvidia-container-runtime-hook: Get hook stage from stdin's 'status' value
Instead of using CRI-O-specific arguments, use the state's REQUIRED
'status' property [1] to ensure we're being called on a created
container (and not a running container or a stopped container).

We could potentially have:

  args := flag.Args()
  if len(args) != 0 {
    flag.Usage()
    os.Exit(2)
  }

but to stay compatible with folks using the hook with older CRI-O
(including the 1.9.x branch), I've just ignored flag.Args() for now.

Also add 'fallthrough' statements [2] to the no-op switch cases, since
that appeared to be the intention.

[1]: https://github.com/opencontainers/runtime-spec/blame/v1.0.1/runtime.md#L16
[2]: https://golang.org/ref/spec#Fallthrough_statements