Remove documentation around the use of new MIG_*_DEVICES envvars
Support for these envvars is being deprecated. The only driver version
that will ever support these envvars is 450.36.04. In lieu of them, we
encourage users to grant the equivalent of docker's `--privileged` flag
for future driver versions in order to configure MIG on a GPU or get
aggregate monitoring information about MIG devices from a GPU.

Signed-off-by: Kevin Klues <kklues@nvidia.com>