Refactor getArgs and fix edge cases
This change refactors the getArgs function that extracts the bundle path
and runc subcommand ("create") from the command line arguments.

The function is changed to accept the command line arguments as a slice
to allow for testing and addresses edge cases with regards to the number
of dashes and an optional equals sign for arguments that are supported by
runc.

Signed-off-by: Evan Lezar <elezar@nvidia.com>