Refactor container runtime to use new packages
This change refactors the NVIDIA container runtime to implement
the oci.Runtime interface. A SyscallExec Runtime is used
to encapsulate the calls to the underlying runtime (runc) and
an oci.SpecModifier encapsulates the processing of the
OCI specification.

Modifications are still implemented in the existing addNVIDIAHooks
function.

The change also includes refactoring / reorganisation of the
functions such as parsing of arguments and adds additional
unit tests.

Signed-off-by: Evan Lezar <elezar@nvidia.com>